+++ resources/views/admin/file/update.php
@@ -13,13 +13,15 @@ use Hirtz\Media\Models\File;
 use Hirtz\Media\Modules\Admin\Controllers\FileController;
 use Hirtz\Media\Modules\Admin\Widgets\Forms\FileActiveForm;
 use Hirtz\Media\Modules\Admin\Widgets\Grids\FileRelationGridContainer;
-use Hirtz\Media\Modules\Admin\Widgets\Grids\TransformationGridView;
 use Hirtz\Media\Modules\Admin\Widgets\Navs\FileHeader;
+use Hirtz\Media\Modules\Admin\Widgets\Navs\FileSubmenu;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Forms\FormContainer;
-use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
 echo FileHeader::make()
+    ->model($file);
+
+echo FileSubmenu::make()
     ->model($file);
 
 echo FormContainer::make()
@@ -29,10 +31,3 @@ echo FormContainer::make()
 
 echo FileRelationGridContainer::make()
     ->file($file);
-
-if ($file->transformation_count) {
-    echo GridContainer::make()
-        ->title(Yii::t('media', 'Transformations'))
-        ->grid(TransformationGridView::make()
-            ->file($file));
-}